Subframe start end messages (#5)
* Created class to deal with CLI. Going to test before future implementations.

* String to integer conversions fixed. Parsing working. File is at [Eclipse's] Debug folder.

* Alterations in lib5grage: use uint8_t to rbStart and ueID. Altered MacConfigRequest.

* Bugfixes when writing to file

* Bugfix: did not close the file after reading

* Implemented procedure to get serialized uplink reservation settings

* First implementation of messages subframe start and End. Needs testing.

* Messages subframeTx start/end syntax correction

* Added subframeRx messages in both UE and BS

The following messages are implemented and working:
- BSSubframeTx.Start
- BSSubframeTx.End
- UESubframeTx.Start
- UESubframeTx.End
- SubframeRx.Start
- SubframeRx.End

There are currently no arguments in this messages, just the string signaling the income or outcome of PDUs.

+++ src/common/lib5grange/lib5grange.h
@@ -254,9 +254,9 @@ namespace lib5grange {
 
      /** Resource allocation configuration struct **/
     typedef struct{
-        uint16_t target_ue_id = ALL_TERMINAL; /**< 16 bit ID of the target terminal **/
-        size_t first_rb = 0;                  /**< First alocated resource block  **/
-        size_t number_of_rb = 132;            /**< Number of allocatced resource blocks  **/
+        uint8_t target_ue_id = ALL_TERMINAL; /**< 16 bit ID of the target terminal **/
+        uint8_t first_rb = 0;                  /**< First alocated resource block  **/
+        uint8_t number_of_rb = 132;            /**< Number of allocatced resource blocks  **/
 
         /** Serializatyion method for the struct**/
         void serialize(vector<uint8_t> & bytes)
